[add] added contact form + old design cleanup
