--- src/Components/Landing/Parallax/ContactForm.scss
@@ -1,142 +1,0 @@
-#ContactForm {
-  display: flex;
-  box-sizing: border-box;
-  flex-flow: column;
-  flex-grow: 1;
-  align-items: stretch;
-  justify-items: stretch;
-  place-content: center;
-  width: 100%;
-  height: 100%;
-  color: #FFFFFF;
-  margin: 0px;
-  font-family: sans-serif;
-  background-color: #414445;
-  text-align: center;
-
-  .sectionTitle {
-    font-weight: lighter;
-    color: #FFFFFF;
-    text-transform: uppercase;
-    font-size: 50px;
-    margin-top: 100px;
-  }
-
-  h4 {
-    width: 100%;
-    text-align: center;
-  }
-
-  form {
-    padding: 20px;
-    background-color: #CEEBFC;
-    border-radius: 30px;
-    align-self: center;
-    align-items: stretch;
-    justify-items: stretch;
-    display: flex;
-    flex-flow: row wrap;
-    @media (max-width: 1024px){
-      width: 80%;
-    }
-    @media (min-width: 1024px){
-      width: 50%;
-    }
-  }
-
-  .userInfoSection {
-    @media (max-width: 800px){
-      width: 100%;
-    }
-    @media (min-width: 800px){
-      width: 50%;
-    }
-  }
-
-  .formSection {
-    padding: 5px;
-    display: flex;
-    flex-direction: row;
-    place-items: center;
-    place-content: center;
-    background-color: #ffffff;
-    border-width: 1px;
-    border-style: solid;
-    border-color: #7CCAFA;
-    color: #3B413C;
-    border-radius: 100px;
-    margin-bottom: 15px;
-    margin-top: 15px;
-    :focus {
-      .icon {
-        color: #ffffff
-      }
-    }
-    @media (min-width: 800px){
-      margin-right: 30px;
-    }
-  }
-
-  .icon {
-    margin-left: 10px;
-    margin-right: 10px;
-    height: 30px;
-    width: 30px;
-  }
-
-  input, textarea {
-    flex-grow: 1;
-    align-self: center;
-    justify-self: center;
-    margin: 0px;
-    padding: 10px;
-    border-style: none;
-    background-color: rgba(0, 0, 0, 0);
-    :focus {
-      outline: none;
-    }
-  }
-
-  .message {
-    background-color: #ffffff;
-    margin-top: 15px;
-    margin-bottom: 15px;
-    border-width: 1px;
-    border-style: solid;
-    border-color: #7CCAFA;
-    color: #3B413C;
-    border-radius: 30px;
-    overflow-wrap: break-word;
-    padding: 20px;
-    justify-self: stretch;
-    align-self: stretch;
-  }
-
-  .submitButton {
-    background-color: rgba(0, 0, 0, 0);
-    text-transform: uppercase;
-    font-size: 20px;
-    padding: 15px;
-    color: #FFFFFF;
-    font-weight: bold;
-    border-radius: 100px;
-    border-width: 3px;
-    border-style: solid;
-    border-color: #FFFFFF;
-    margin-top: 30px;
-    margin-bottom: 100px;
-    :hover {
-      background-color: #CEEBFC;
-      border-color: #CEEBFC;
-      color: #3B413C;
-      transition: all 0.5s ease-out;
-      box-shadow: -0px -0px 12px 2px rgba(0, 0, 0, 1);
-    }
-    @media (max-width: 1024px){
-      width: 40%;
-    }
-    @media (min-width: 1024px){
-      width: 15%;
-    }
-  }
-}
