--- src/Components/Landing/Parallax/Parallax.scss
@@ -1,179 +1,0 @@
-.Parallax {
-  display: flex;
-  flex-direction: column;
-  box-sizing: border-box;
-  font-family: 'Barlow Semi Condesed', 'Roboto', sans-serif;
-  place-items: center;
-  place-content: center;
-  flex-grow: 1;
-  background-color: #CEEBFC;
-  padding: 0;
-  margin: 0;
-  position: relative;
-  left: 0;
-  right: 0;
-  height: 100%;
-  color: #171B1C;
-
-  .sectionTitle {
-    font-family: 'Barlow Semi Condensed', 'Roboto', sans-serif;
-    font-weight: lighter;
-    color: #171B1C;
-    text-transform: uppercase;
-    font-size: 50px;
-    margin-top: 100px;
-  }
-
-  p {
-    @media (max-width: 1024px){
-      width: 75%;
-    }
-    @media (min-width: 1024px){
-      width: 45%;
-    }
-    font-size: 18px;
-    text-align: center;
-  }
-}
-
-.MainPres {
-  display: flex;
-  flex-direction: column;
-  place-items: center;
-  place-content: center;
-  height: 100vh;
-  width: 100%;
-  color: #ffffff;
-  background-color: #C4E7FC;
-  background-size: cover;
-  margin: 0;
-  font-family: 'Barlow Semi Condensed', 'Roboto', sans-serif;
-  text-shadow: #000000 1px 0 10px;
-  font-weight: normal;
-
-  h1 {
-    color: #ffffff;
-    text-transform: uppercase;
-    font-weight: lighter;
-    text-align: center;
-    @media (max-width: 1024px){
-      font-size: 60px;
-    }
-    @media (min-width: 1024px){
-      font-size: 100px;
-    }
-  }
-
-  h3 {
-    color: #ffffff;
-    font-size: 30px;
-    font-weight: lighter;
-    text-transform: none;
-    text-align: center;
-  }
-
-  .knowMore {
-    background-color: rgba(0, 0, 0, 0);
-    text-transform: uppercase;
-    text-decoration: none;
-    font-size: 25px;
-    padding: 20px;
-    color: #FFFFFF;
-    font-weight: bold;
-    border-radius: 30px;
-    border-width: 3px;
-    border-style: solid;
-    border-color: #FFFFFF;
-    margin-top: 30px;
-    margin-bottom: 100px;
-    :hover {
-      background-color: #CEEBFC;
-      border-color: #CEEBFC;
-      color: #3B413C;
-      transition: all 0.5s ease-out;
-      box-shadow: -0px -0px 12px 2px rgba(0, 0, 0, 1);
-    }
-    :focus {
-      outline: none;
-    }
-  }
-}
-
-.Description {
-  display: flex;
-  box-sizing: border-box;
-  flex-flow: row wrap;
-  flex-grow: 1;
-  flex-shrink: 1;
-  align-items: stretch;
-  justify-items: stretch;
-  place-content: center;
-  width: 90%;
-  height: 100%;
-  color: #3B413C;
-  margin: 0;
-  font-family: sans-serif;
-  text-align: center !important;
-
-  .overlay {
-    background-color: #7CCAFA;
-  }
-
-  .box {
-    text-align: center !important;
-    display: flex;
-    flex-grow: 1;
-    place-items: center;
-    place-content: center;
-    flex-direction: column;
-    background-color: #ffffff;
-    margin: 20px;
-    padding: 30px;
-    overflow-y: auto;
-    justify-self: stretch;
-    align-self: stretch;
-    box-sizing: border-box;
-    margin-bottom: 100px;
-    border-style: none;
-    @media (max-width: 1024px){
-      width: 100%;
-    }
-    @media (min-width: 1024px){
-      width: 15%;
-    }
-    :hover {
-      transition: all 0.3s ease-out;
-      box-shadow: -0px -0px 12px 5px rgba(0, 0, 0, 0.3);
-      background-color: #7CCAFA;
-      .icon {
-        color: #FFFFFF;
-      }
-    }
-    p {
-      text-align: center;
-    }
-  }
-
-  .icon {
-    align-self: center;
-    justify-self: center;
-    text-align: center;
-    height: 15%;
-    width: 15%;
-    color: #5FBFF9;
-  }
-
-
-  h4 {
-    width: 100%;
-    text-align: center;
-    font-weight: bold;
-  }
-
-  p {
-    width: 100%;
-    text-align: center;
-    font-size: 16px;
-  }
-
-}
