--- src/Components/Landing/Parallax/PreviewSection.scss
@@ -1,125 +1,0 @@
-.PreviewSection {
-  display: flex;
-  box-sizing: border-box;
-  flex-flow: column;
-  flex-grow: 1;
-  align-items: stretch;
-  justify-items: stretch;
-  place-content: center;
-  width: 100%;
-  height: 100%;
-  margin: 0px;
-  font-family: sans-serif;
-  background-color: #99D6FB;
-  text-align: center;
-
-  p {
-    align-self: center;
-    width: 60%;
-  }
-
-  h4 {
-    text-transform: uppercase;
-    font-family: 'Barlow Semi Condensed', 'Roboto', sans-serif;
-    font-weight: bold;
-    font-size: 25px;
-    background-color: #C4E7FC;
-    border-radius: 10px;
-    color: #2C2F30;
-    padding: 10px;
-    margin-top: 20px;
-    margin-bottom: 20px;
-  }
-
-  .previewPic {
-    display: flex;
-    flex-grow: 1;
-    height: 100%;
-    place-items: center;
-    place-content: center;
-    align-self: center;
-    justify-self: center;
-    img {
-      width: 100%;
-      align-self: center;
-      justify-self: center;
-    }
-    @media (max-width: 1024px){
-      width: 100%;
-    }
-    @media (min-width: 1024px){
-      width: 30%;
-    }
-  }
-
-  .previewText {
-    justify-self: center;
-    align-self: center;
-    padding: 20px;
-    div {
-      border-radius: 20px;
-      background-color: #E9F6FD;
-      :hover {
-        background-color: #C4E7FC;
-      }
-    }
-    p {
-      display: flex;
-      width: 100%;
-      align-self: center;
-      justify-self: center;
-      text-align: center;
-      font-size: 16px
-    }
-    @media (max-width: 1024px){
-      width: 100%;
-      div {
-        padding: 20px;
-      }
-    }
-    @media (min-width: 1024px){
-      width: 60%;
-      div {
-        padding: 50px;
-      }
-    }
-    :hover {
-      h4 {
-        box-shadow: -0px -0px 8px 3px rgba(0, 0, 0, 0.3);
-        background-color: #57AEE3;
-      }
-    }
-  }
-
-  .previewContainer {
-    padding: 30px;
-    margin: 20px;
-    margin-bottom: 100px;
-    text-align: center;
-    display: flex;
-    flex-grow: 1;
-    place-items: center;
-    place-content: center;
-    flex-flow: row wrap;
-    background-color: #FFFFFF;
-    overflow-y: auto;
-    justify-self: center;
-    align-self: center;
-    border-radius: 30px;
-    box-sizing: border-box;
-    box-shadow: -0px -0px 12px 5px rgba(0, 0, 0, 0.3);
-    @media (max-width: 1024px){
-      width: 90%;
-    }
-    @media (min-width: 1024px){
-      width: 60%;
-    }
-  }
-
-  .sectionTitle {
-    font-weight: lighter;
-    text-transform: uppercase;
-    font-size: 50px;
-    margin-top: 100px;
-  }
-}
